feat: connect API calls to save button in CreateEditSite
